Revert "Issue 6183: Change readSegment method Parameter from int to long in SegmentHelper (#6239)"
This reverts commit a3d957497c2b3af0f5ec25be6d1652a26368a006.

+++ cli/admin/src/main/java/io/pravega/cli/admin/segmentstore/ReadSegmentRangeCommand.java
@@ -111,8 +111,10 @@ public class ReadSegmentRangeCommand extends SegmentStoreCommand {
         FileOutputStream fileOutputStream = new FileOutputStream(file, true);
         while (bytesToRead > 0) {
             long bufferLength = Math.min(READ_WRITE_BUFFER_SIZE, bytesToRead);
+            // TODO: readSegment should take in longs as described here: https://github.com/pravega/pravega/issues/6183
+            // TODO: Remove casting to int once 6183 is resolved.
             CompletableFuture<WireCommands.SegmentRead> reply = segmentHelper.readSegment(fullyQualifiedSegmentName,
-                    currentOffset, (int) bufferLength, new PravegaNodeUri(segmentStoreHost, getServiceConfig().getAdminGatewayPort()), super.authHelper.retrieveMasterToken());
+                    (int) currentOffset, (int) bufferLength, new PravegaNodeUri(segmentStoreHost, getServiceConfig().getAdminGatewayPort()), super.authHelper.retrieveMasterToken());
             WireCommands.SegmentRead bufferRead = reply.get(REQUEST_TIMEOUT_SECONDS, TimeUnit.SECONDS);
             int bytesRead = bufferRead.getData().readableBytes();
             // Write the buffer into the file.
